pyproject.toml
[project]
name = "home-assignment-ai-engineer"
version = "0.1.0"
description = "AI Engineer Home Assignment - Async FastAPI Service"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "httpx>=0.25.0",
    "python-multipart>=0.0.6",
    "diagrams>=0.23.0",
    "google-generativeai>=0.3.0",
    "mcp>=0.5.0",
    "fastmcp>=0.6.0",
    "pillow>=10.0.0",
    "graphviz>=0.20.0",
    "python-dotenv>=1.0.0",
    "sse-starlette>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.black]
line-length = 88

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.2.1",
]
